fix don't select network by default if not filtering
